user-wallet-service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg-binary==3.2.9   
python-decouple==3.8     
requests==2.31.0